bot/requirements.txt
python-telegram-bot[ext]==22.1
supabase==2.13.0
python-dotenv==1.1.0
